RESUME ATS TRACKING GEN AI APP/requirements.txt
streamlit 
python-dotenv
google-generativeai
pdf2image